OBG-859. Make checkstyle happy after renaming classes
