+++ opba-api-security-signer-generator/opba-api-security-signer-generator-impl/src/main/java/de/adorsys/opba/api/security/generator/normalizer/DataToSignProviderGenerator.java
@@ -1,4 +1,4 @@
-package de.adorsys.opba.api.security.generator.signer;
+package de.adorsys.opba.api.security.generator.normalizer;
 
 import com.google.common.base.Strings;
 import com.squareup.javapoet.AnnotationSpec;
@@ -9,11 +9,11 @@ import com.squareup.javapoet.JavaFile;
 import com.squareup.javapoet.MethodSpec;
 import com.squareup.javapoet.ParameterSpec;
 import com.squareup.javapoet.TypeSpec;
+import de.adorsys.opba.api.security.generator.api.DataToSignProvider;
 import de.adorsys.opba.api.security.generator.api.GeneratedDataToSignNormalizer;
 import de.adorsys.opba.api.security.generator.api.MatcherUtil;
 import de.adorsys.opba.api.security.generator.api.RequestDataToSignNormalizer;
 import de.adorsys.opba.api.security.generator.api.RequestToSign;
-import de.adorsys.opba.api.security.generator.normalizer.RequestDataToSignNormalizerGenerator;
 import io.swagger.v3.oas.models.Operation;
 
 import javax.annotation.Generated;
@@ -24,7 +24,8 @@ import java.io.IOException;
 import java.util.Map;
 
 public class DataToSignProviderGenerator {
-    private static final String CLASS_PURPOSE_COMMENT = "This class provides request signature canonicalization classes (convert Request to String to sign)";
+    private static final String CLASS_PURPOSE_COMMENT =
+            "This class provides request signature canonicalization classes (convert Request to String to sign)";
 
     private final RequestDataToSignNormalizerGenerator requestDataToSignNormalizerGenerator;
 
@@ -32,7 +33,7 @@ public class DataToSignProviderGenerator {
         this.requestDataToSignNormalizerGenerator = dataToSignProvider;
     }
 
-    public void generate(String packageName, GeneratedDataToSignNormalizer signerConfig, Filer filer, Map<String, Map<de.adorsys.opba.api.security.generator.api.DataToSignProvider.HttpMethod, Operation>> requestSpecConfig) {
+    public void generate(String packageName, GeneratedDataToSignNormalizer signerConfig, Filer filer, Map<String, Map<DataToSignProvider.HttpMethod, Operation>> requestSpecConfig) {
         ClassName targetClass = ClassName.get(
                 packageName,
                 Strings.isNullOrEmpty(signerConfig.signerClassName()) ? "RequestSignerImpl" : signerConfig.signerClassName()
@@ -104,14 +105,14 @@ public class DataToSignProviderGenerator {
     }
 
     private void implementSignerInterface(TypeSpec.Builder signer) {
-        signer.addSuperinterface(ClassName.get(de.adorsys.opba.api.security.generator.api.DataToSignProvider.class));
+        signer.addSuperinterface(ClassName.get(DataToSignProvider.class));
     }
 
     private void addWithBasePathMethod(TypeSpec.Builder signer, ClassName targetClass) {
         MethodSpec.Builder method = MethodSpec
                 .methodBuilder("withBasePath")
                 .addAnnotation(ClassName.get(Override.class))
-                .returns(ClassName.get(de.adorsys.opba.api.security.generator.api.DataToSignProvider.class))
+                .returns(ClassName.get(DataToSignProvider.class))
                 .addModifiers(Modifier.PUBLIC);
 
         ParameterSpec basePath = ParameterSpec.builder(ClassName.get(String.class), "basePath").build();
@@ -132,7 +133,7 @@ public class DataToSignProviderGenerator {
             ClassName targetClass,
             GeneratedDataToSignNormalizer signerConfig,
             Filer filer,
-            Map<String, Map<de.adorsys.opba.api.security.generator.api.DataToSignProvider.HttpMethod, Operation>> requestSpecConfig
+            Map<String, Map<DataToSignProvider.HttpMethod, Operation>> requestSpecConfig
     ) {
         MethodSpec.Builder method = MethodSpec
                 .methodBuilder("normalizerFor")
@@ -166,11 +167,11 @@ public class DataToSignProviderGenerator {
         signer.addMethod(method.build());
     }
 
-    private void returnSignerIfBlock(ParameterSpec toSign, CodeBlock.Builder block, String path, de.adorsys.opba.api.security.generator.api.DataToSignProvider.HttpMethod requestMethod, ClassName className) {
+    private void returnSignerIfBlock(ParameterSpec toSign, CodeBlock.Builder block, String path, DataToSignProvider.HttpMethod requestMethod, ClassName className) {
         CodeBlock ifBlock = CodeBlock.builder()
                 .beginControlFlow(
                         "if ($T.$L == $N.getMethod() && $T.matches($S, computedPath))",
-                        de.adorsys.opba.api.security.generator.api.DataToSignProvider.HttpMethod.class,
+                        DataToSignProvider.HttpMethod.class,
                         requestMethod,
                         toSign,
                         ClassName.get(MatcherUtil.class),
