+++ opba-api-security-signer-generator/opba-api-security-signer-generator-impl/src/test/resources/expected/ExpectedDataToSignProvider.java
@@ -9,7 +9,7 @@ import java.lang.String;
 import javax.annotation.Generated;
 
 @Generated(
-        value = "de.adorsys.opba.api.security.generator.signer.DataToSignProviderGenerator",
+        value = "de.adorsys.opba.api.security.generator.normalizer.DataToSignProviderGenerator",
         comments = "This class provides request signature canonicalization classes (convert Request to String to sign)"
 )
 public class TestSigner implements DataToSignProvider {
